code cleanup and refactoring
