+++ examples/depricated/omop_etl_vista_config_templated.json
