+++ examples/depricated/omop_etl_vista_omop_concepts.json
@@ -70,7 +70,7 @@
         "visit_occurrence": {
             "time_start": "@visit_start_datetime || @visit_start_date",
             "time_end": "@visit_end_datetime || @visit_end_date",
-            "code": "@visit_concept_id",
+            "code": "$omop:@visit_concept_id || @visit_source_value",
             "properties": [
                 {
                     "name": "table",
@@ -91,18 +91,13 @@
                     "name": "care_site_id",
                     "value": "@care_site_id",
                     "type": "int"
-                },
-                {
-                    "name": "visit_source_concept_id",
-                    "value": "@visit_source_concept_id",
-                    "type": "string"
                 }
             ]
         },
         "visit_detail": {
             "time_start": "@visit_detail_start_datetime || @visit_detail_start_date",
             "time_end": "@visit_detail_end_datetime || @visit_detail_end_date",
-            "code": "STANFORD_VISIT_DETAIL/{@visit_detail_source_value | split('|', 3, 'Unknown')}",
+            "code": "$omop:@visit_detail_concept_id || STANFORD_VISIT_DETAIL/{@visit_detail_source_value | split('|', 3, 'Unknown')}",
             "properties": [
                 {
                     "name": "table",
@@ -129,7 +124,7 @@
         "condition_occurrence": {
             "time_start": "@condition_start_datetime || @condition_start_date",
             "time_end": "@condition_end_datetime || @condition_end_date",
-            "code": "@condition_concept_id",
+            "code": "$omop:@condition_source_concept_id",
             "properties": [
                 {
                     "name": "table",
@@ -151,7 +146,7 @@
         "drug_exposure": {
             "time_start": "@drug_exposure_start_datetime || @drug_exposure_start_date",
             "time_end": "@drug_exposure_end_datetime || @drug_exposure_end_date",
-            "code": "@drug_concept_id",
+            "code": "$omop:@drug_source_concept_id",
             "properties": [
                 {
                     "name": "table",
@@ -172,7 +167,7 @@
         },
         "procedure_occurrence": {
             "time_start": "@procedure_datetime || @procedure_date",
-            "code": "@procedure_concept_id",
+            "code": "$omop:@procedure_source_concept_id",
             "properties": [
                 {
                     "name": "table",
@@ -204,11 +199,21 @@
         "device_exposure": {
             "time_start": "@device_exposure_start_datetime || @device_exposure_start_date",
             "time_end": "@device_exposure_end_datetime || @device_exposure_end_date",
-            "code": "@device_concept_id",
+            "code": "$omop:@device_concept_id",
             "properties": [
                 {
                     "name": "table",
                     "value": "device_exposure",
+                    "type": "string"
+                },
+                {
+                    "name": "source_code",
+                    "value": "$omop:@device_source_concept_id",
+                    "type": "string"
+                },
+                {
+                    "name": "source_value",
+                    "value": "@device_source_value",
                     "type": "string"
                 },
                 {
@@ -230,7 +235,7 @@
         },
         "measurement": {
             "time_start": "@measurement_datetime || @measurement_date",
-            "code": "@measurement_concept_id",
+            "code": "$omop:@measurement_source_concept_id",
             "numeric_value": "@value_as_number",
             "text_value": "@value_source_value",
             "properties": [
@@ -258,7 +263,7 @@
         },
         "observation": {
             "time_start": "@observation_datetime || @observation_date",
-            "code": "@observation_concept_id",
+            "code": "$omop:@observation_source_concept_id",
             "numeric_value": "@value_as_number",
             "text_value": "@value_as_string",
             "properties": [
