--- scripts/analyze_config_mapping.py
@@ -1,200 +1,0 @@
-"""
-Analyze how omop_etl_vista_config.json creates different codes than omop.py
-"""
-
-import json
-from pathlib import Path
-
-
-def analyze_config_code_mapping():
-    """
-    Analyze the config file to understand code generation.
-    """
-    print("=" * 80)
-    print("CONFIG-BASED CODE MAPPING ANALYSIS")
-    print("=" * 80)
-
-    config_path = Path("omop_etl_vista_config.json")
-    with open(config_path) as f:
-        config = json.load(f)
-
-    # ========================================================================
-    # 1. CANONICAL EVENTS
-    # ========================================================================
-    print("\n[1] CANONICAL EVENTS (Fixed Codes)")
-    print("-" * 80)
-
-    canonical = config.get("canonical_events", {})
-    print(f"\nFound {len(canonical)} canonical events:")
-
-    for event_name, event_config in canonical.items():
-        table = event_config.get("table")
-        code = event_config.get("code", f"MEDS_{event_name.upper()}")
-        print(f"  {event_name:20} -> {code:30} (from {table})")
-
-    # ========================================================================
-    # 2. TEMPLATE-BASED CODE MAPPINGS
-    # ========================================================================
-    print("\n[2] TEMPLATE-BASED CODE MAPPINGS")
-    print("-" * 80)
-
-    print("\nTables using templates for code generation:")
-
-    for table_name, table_config in config.get("tables", {}).items():
-        code_mappings = table_config.get("code_mappings", {})
-
-        has_template = False
-
-        # Check source_value for templates
-        if "source_value" in code_mappings:
-            sv_config = code_mappings["source_value"]
-            if isinstance(sv_config, dict) and "template" in sv_config:
-                template = sv_config["template"]
-                transforms = sv_config.get("transforms", [])
-                print(f"\n  {table_name} (source_value):")
-                print(f"    Template: {template}")
-                if transforms:
-                    print(f"    Transforms: {transforms}")
-                has_template = True
-
-        # Check concept_id for templates
-        if "concept_id" in code_mappings:
-            ci_config = code_mappings["concept_id"]
-            if isinstance(ci_config, dict) and "template" in ci_config:
-                template = ci_config["template"]
-                transforms = ci_config.get("transforms", [])
-                print(f"\n  {table_name} (concept_id):")
-                print(f"    Template: {template}")
-                if transforms:
-                    print(f"    Transforms: {transforms}")
-                has_template = True
-
-        # Check for field-based mappings (no template)
-        if not has_template and code_mappings:
-            for mapping_type, mapping_config in code_mappings.items():
-                if isinstance(mapping_config, dict):
-                    if "field" in mapping_config:
-                        print(f"\n  {table_name} ({mapping_type}):")
-                        print(f"    Field: {mapping_config['field']}")
-                    elif "concept_id_field" in mapping_config:
-                        print(f"\n  {table_name} ({mapping_type}):")
-                        print(f"    Concept lookup: {mapping_config['concept_id_field']}")
-
-    # ========================================================================
-    # 3. DEMOGRAPHIC CODES
-    # ========================================================================
-    print("\n[3] DEMOGRAPHIC/PERSON TABLE")
-    print("-" * 80)
-
-    if "person" in config.get("canonical_events", {}):
-        person_config = config["canonical_events"]["person"]
-        print("\nPerson table handling:")
-        print(f"  Source table: {person_config.get('table')}")
-        print(f"  Birth code: {person_config.get('code', 'MEDS_BIRTH')}")
-
-        # Check for death
-        if "death" in config.get("canonical_events", {}):
-            death_config = config["canonical_events"]["death"]
-            print(f"  Death code: {death_config.get('code', 'MEDS_DEATH')}")
-
-    # ========================================================================
-    # 4. IMAGE CODES
-    # ========================================================================
-    print("\n[4] IMAGE/IMAGING CODES")
-    print("-" * 80)
-
-    # Check image table
-    if "image" in config.get("tables", {}):
-        image_config = config["tables"]["image"]
-        print("\nImage table found:")
-
-        code_mappings = image_config.get("code_mappings", {})
-        for mapping_type, mapping_config in code_mappings.items():
-            if isinstance(mapping_config, dict) and "template" in mapping_config:
-                print(f"  Mapping type: {mapping_type}")
-                print(f"  Template: {mapping_config['template']}")
-                print("  This creates codes like: IMAGE/<modality>|<anatomic_site>")
-
-    # Check imaging table
-    if "imaging" in config.get("tables", {}):
-        imaging_config = config["tables"]["imaging"]
-        print("\nImaging table found:")
-
-        code_mappings = imaging_config.get("code_mappings", {})
-        for mapping_type, mapping_config in code_mappings.items():
-            if isinstance(mapping_config, dict) and "template" in mapping_config:
-                print(f"  Mapping type: {mapping_type}")
-                print(f"  Template: {mapping_config['template']}")
-
-    # ========================================================================
-    # 5. NOTE CODES
-    # ========================================================================
-    print("\n[5] NOTE CODES")
-    print("-" * 80)
-
-    if "note" in config.get("tables", {}):
-        note_config = config["tables"]["note"]
-        print("\nNote table found:")
-
-        code_mappings = note_config.get("code_mappings", {})
-        for mapping_type, mapping_config in code_mappings.items():
-            if isinstance(mapping_config, dict) and "template" in mapping_config:
-                print(f"  Mapping type: {mapping_type}")
-                print(f"  Template: {mapping_config['template']}")
-                print("  This creates codes like: NOTE/<note_type>")
-
-    # ========================================================================
-    # 6. COMPARISON WITH OMOP.PY HARDCODED LOGIC
-    # ========================================================================
-    print("\n[6] KEY DIFFERENCES FROM OMOP.PY")
-    print("-" * 80)
-
-    print("\nomop.py (hardcoded):")
-    print("  - Uses raw concept IDs (e.g., '4180938')")
-    print("  - Simple demographic codes: 'Gender/M', 'Gender/F', 'Race/2'")
-    print("  - Domain-based codes: 'Domain/OMOP generated'")
-    print("  - Limited custom Stanford codes")
-
-    print("\nomop_streaming.py (config-driven):")
-    print("  - Uses template-based code generation")
-    print("  - Structured imaging codes: 'IMAGE/<modality>|<anatomic_site>'")
-    print("  - Structured note codes: 'NOTE/<note_type>'")
-    print("  - Maps concept IDs through concept table OR templates")
-    print("  - More flexible metadata preservation")
-
-    print("\nImplications:")
-    print("  1. Config-based approach creates MORE unique codes (48,260 vs 46,913)")
-    print("  2. Config generates structured, hierarchical codes")
-    print("  3. Config preserves more OMOP metadata (51 vs 22 columns)")
-    print("  4. Different filtering logic leads to ~379K row difference")
-
-    # ========================================================================
-    # 7. EXPLAIN THE 379K ROW DIFFERENCE
-    # ========================================================================
-    print("\n[7] EXPLAINING THE 379K ROW DIFFERENCE")
-    print("-" * 80)
-
-    print("\nFrom the comparison analysis:")
-    print("  - omop.py has ~6.7M rows with 43 unique codes")
-    print("  - omop_streaming.py has ~6.3M rows with 1,390 unique codes")
-    print("  - Net difference: ~379K rows (0.20%)")
-
-    print("\nThe swap is happening because:")
-    print("  1. omop.py includes many instances of concept ID '4180938'")
-    print("  2. omop.py includes demographic codes (Gender/Race/Ethnicity)")
-    print("  3. omop_streaming.py generates detailed NOTE/* codes")
-    print("  4. omop_streaming.py generates detailed IMAGE/* codes")
-
-    print("\nThis is NOT a bug - it's different semantic choices:")
-    print("  - omop.py: More compact, fewer code variations")
-    print("  - omop_streaming.py: More detailed, preserves data structure")
-
-    print("\nRecommendation:")
-    print("  ✅ Use omop_streaming.py for richer, more structured MEDS data")
-    print("  ✅ The config approach is more maintainable and flexible")
-    print("  ✅ Better for downstream analysis with hierarchical codes")
-    print()
-
-
-if __name__ == "__main__":
-    analyze_config_code_mapping()
